refactor: [#336] move ServiceEndpoint to shared module
ServiceEndpoint is a layer-agnostic validated wrapper (like DomainName,
BindAddress) used by both application and infrastructure layers.

Moving it from infrastructure/external_validators to shared/ makes
its role clearer as a generic utility type that doesn't belong to
any specific layer.
